Add NodeOptions constructor to Cell (#13)
* Add NodeOptions constructor to Cell

* Add constructor declaration and
definition for Cell to take a
NodeOptions structure and parse
the name and period. This is
so that the Cell can be configured
as a component.

* Add utilities header to make parsing
NodeOptions more convenient. This utility
can be used by derived classes outside
of modulo core.

* Pass NodeOptions to lifecycle node

* In Cell constructor taking a NodeOptions,
pass the whole options structure through
to the base LifecycleNode so that everything
(namespace remapping, parameters, internal
comms, etc) are preserved

* Improve docstrings