Parameter translation improvements (#57)
* Add const parameter reader

* Add read parameter translator that takes a reference to a
const ParameterInterface. This is useful for validating parameters
without modifying them (in case of invalidity), where an original
parameter object is needed as a reference (e.g. for matrix size,
state compatibility or other)

* Refactor and split tests by parameter type

* Rename message translators

* Rename ReadStateConversion and WriteStateConversion to read_message
and write_message respectively. For consistency in translators,
lower_snake_case is used for source files that have no specific class.

* Rename test file and update references

* Remove unused variable to suppress warning

* Touchup test case for constness

* Ensure that the const reader makes a new parameter interface
and cannot modify the original parameter through the "copy"

* Copy parameter helper

* Implement a parameter copy helper for ParameterInterface pointers
so that a referenced value can be updated without breaking the reference

* Add IncompatibleParameterException to resolve some TODOs

* Extend tests

* Update constness

* Mark copy method source parameter as const

* Enable matrix translation and test