User-defined callback on inputs (C++) (#124)
* Insert a user callback into SubscriptionHandler

* Add private class property and setter for user_callback_,
a void function

* Invoke user callback within the subscription callback, after
reading the message

* Move callback error handling to helper function to simplify
implementation and reduce code duplication

* Allow setting user callback directly when getting the
subscription callback with a function overload

* Implement user callback in component add_input

* Resolve TODO by overloading add_input with a user_callback
argument after the shared data pointer. This allows a void function
to be executed after the message is read into the data (as opposed
to the generic input callback which takes a raw message). The
default add_input with a data pointer passes an empty user callback
instead.

* Update tests for adding inputs with user callback

* Add tests on component level

* Adjust docstrings

Co-authored-by: Dominic Reber <dominic.reber@epfl.ch>